+++ CHANGES.md
@@ -5,6 +5,10 @@
 ### Fixes
 - Invalid option names in pyproject.toml file will now stop Robotidy (with optional hint message) [#150](https://github.com/MarketSquare/robotframework-tidy/issues/150)
 
+### Features
+- Allow to use spaces in pyproject.toml configuration file [#148](https://github.com/MarketSquare/robotframework-tidy/issues/148)
+
+
 ## 1.3.0
 
 ### Transformers